requirements.txt
aiohttp==3.7.1
aiohttp_cors==0.7.0
astral==1.10.1
async_timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
certifi>=2020.6.20
ciso8601==2.1.3
httpx==0.16.1
importlib-metadata==1.6.0;python_version<'3.8'
jinja2>=2.11.2
PyJWT==1.7.1
cryptography==3.2
pip>=8.0.3
python-slugify==4.0.1
pytz>=2020.1
pyyaml==5.3.1
requests==2.25.0
ruamel.yaml==0.15.100
sqlalchemy==1.3.20
voluptuous==0.12.0
voluptuous-serialize==2.4.0
yarl==1.4.2
colorlog>=4.0.0

# Constrain urllib3 to ensure we deal with CVE-2019-11236 & CVE-2019-11324
urllib3>=1.24.3

# Constrain httplib2 to protect against CVE-2020-11078
httplib2>=0.18.0
